Moved docker lib to it's own module.
