+++ docker/git.go
@@ -1,9 +1,11 @@
-package cali
+package docker
 
 import (
 	"crypto/md5"
 	"fmt"
 	"os"
+
+	"path"
 
 	log "github.com/Sirupsen/logrus"
 	"github.com/docker/docker/api/types/container"
@@ -12,7 +14,7 @@ import (
 
 // GitCheckoutConfig is input for Git.Checkout
 type GitCheckoutConfig struct {
-	Repo, Branch, RelPath, Image string
+	Repo, Branch, RelPath, Image, WorkDir string
 }
 
 const gitImage = "indiehosters/git:latest"
@@ -26,6 +28,41 @@ func (c *DockerClient) Git() *Git {
 type Git struct {
 	c     *DockerClient
 	Image string
+}
+
+// BindFromGit creates a data container with a git clone inside and mounts its volumes inside your app container
+// If there is no valid Git repo set in config, the noGit callback function will be executed instead
+func (c *DockerClient) BindFromGit(cfg *GitCheckoutConfig, noGit func() error) error {
+	cli := NewClient()
+	cli.Host = c.Host
+	cli.Registry = c.Registry
+
+	if err := cli.InitDocker(); err != nil {
+		return err
+	}
+
+	if cfg.Repo != "" {
+		// Build code from data volume
+		git := cli.Git()
+
+		if cfg.Image != "" {
+			git.Image = cfg.Image
+		}
+		id, err := git.Checkout(cfg)
+
+		if err != nil {
+			return err
+		}
+		c.HostConf.VolumesFrom = []string{id}
+
+		if cfg.RelPath != "" {
+			c.SetWorkDir(path.Join(cfg.WorkDir, cfg.RelPath))
+		}
+	} else {
+		// Execute callback
+		noGit()
+	}
+	return nil
 }
 
 // Checkout will create and start a container, checkout repo and leave container stopped
@@ -69,7 +106,7 @@ func (g *Git) Checkout(cfg *GitCheckoutConfig) (string, error) {
 	g.c.SetHostConf(&hc)
 	g.c.SetNetConf(&nc)
 
-	id, err := g.c.StartContainer(false, name)
+	id, err := g.c.ExecContainer(false, name)
 
 	if err != nil {
 		return "", fmt.Errorf("Failed to create data container for %s: %s", cfg.Repo, err)
@@ -100,5 +137,5 @@ func (g *Git) Pull(name string) (string, error) {
 	g.c.SetHostConf(&hc)
 	g.c.SetNetConf(&nc)
 
-	return g.c.StartContainer(true, "")
+	return g.c.ExecContainer(true, "")
 }
