+++ task.go
@@ -2,17 +2,19 @@ package cali
 
 import (
 	"fmt"
+	"log"
 	"os/user"
 	"path/filepath"
 	"runtime"
 	"strings"
-	"log"
+
+	"github.com/edupo/cali/docker"
 )
 
 // Task is the action performed when it's parent command is run
 type Task struct {
 	f, init TaskFunc
-	*DockerClient
+	*docker.DockerClient
 }
 
 // TaskFunc is a function executed by a Task when the command the Task belongs to is run
@@ -27,13 +29,13 @@ var defaultTaskFunc TaskFunc = func(t *Task, args []string) {
 	if err := t.InitDocker(); err != nil {
 		log.Fatalf("Error initialising Docker: %s", err)
 	}
-	if _, err := t.StartContainer(true, ""); err != nil {
+	if _, err := t.ExecContainer(true, ""); err != nil {
 		log.Fatalf("Error executing task: %s", err)
 	}
 }
 
 func NewTask() *Task {
-	return &Task{DockerClient: NewDockerClient()}
+	return &Task{DockerClient: docker.NewClient()}
 }
 
 // SetFunc sets the TaskFunc which is run when the parent command is run
@@ -57,6 +59,7 @@ func (t *Task) SetInitFunc(f TaskFunc) {
 func (t *Task) SetDefaults(args []string) error {
 	t.SetWorkDir(workDir)
 	t.SetRegistry(flags.GetString("docker-registry"))
+	t.Host = flags.GetString("docker-host")
 	awsDir, err := t.Bind("~/.aws", "/root/.aws")
 	if err != nil {
 		return err
